refactor(webui): split config service helpers
